% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignReactionConditions.R
\name{assignReactionConditions}
\alias{assignReactionConditions}
\title{Assign reaction conditions to students}
\usage{
assignReactionConditions()
}
\value{
No return value. Called to generate \code{students_rxn_params.rda}
}
\description{
Loads variables, student names, and enzyme properties, then assigns reaction conditions and saves the results.
}
\details{
Creates timestamped subfolders within \verb{output/} directory for variables, student data, and answers.
Assigns enzyme substrate and inhibition types to each student (student always receives \code{no_inhibition} and then one of: \code{competitive}, \code{non-competitive}, and \code{uncompetitive}).
Prepares \code{student_rxn_params.rda} within \code{Variables} subfolder for downstream analysis. \code{Variables} subfolder is of little importance to the user.
}
\seealso{
\code{\link{runMaudr}}
}
