% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculateGradient}
\alias{calculateGradient}
\title{Calculate absorbance-time gradient
Calculates absorbance-time gradient for given parameters}
\usage{
calculateGradient(Vmax, Km, substrate_conc, enz_vol, cuv_vol, eps)
}
\value{
Calculated absorbance-time gradient when called in \code{processReactionParameters} and \code{generateAbsVsTimeData}. User does not need to interact.
}
\description{
Calculate absorbance-time gradient
Calculates absorbance-time gradient for given parameters
}
\details{
Internally calls \code{calculateReactionRate} to simulate absorbance changes and spectrophotometer assays
}
\seealso{
\code{\link{processReactionParameters}} \code{\link{generateAbsVsTime}} \code{\link{calculateReactionRate}}
}
