% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculateReactionRate}
\alias{calculateReactionRate}
\title{Calculate reaction rate with Michaelis-Menten equation
Calculates the reaction rate for given parameters and substrate concentration}
\usage{
calculateReactionRate(Vmax, Km, substrate_conc)
}
\value{
Calculated reaction rate when called in \code{processReactionParameters} and \code{generateAbsVsTimeData}. User does not need to interact.
}
\description{
Calculate reaction rate with Michaelis-Menten equation
Calculates the reaction rate for given parameters and substrate concentration
}
\details{
Uses Michaelis-Menten \eqn{v = (Vmax * [S]) / (Km + [S])} to calculate reaction rate.
Adds jitter to concentration levels greater than 0 to simulate wet lab variability.
}
\seealso{
\code{\link{processReactionParameters}} \code{\link{generateAbsVsTime}}
}
