% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{enzyme_properties}
\alias{enzyme_properties}
\title{Example enzyme properties}
\format{
A data frame with variables:
\describe{
\item{rxn_substrate}{Name of alcohol substrate (e.g. "ethanol", "propanol")}
\item{Kcat}{Turnover number (numeric)}
\item{Km}{Michaelis constant (numeric, mM)}
\item{Vmax}{Maximum reaction rate (numeric)}
\item{enzyme_conc}{Enzyme concentration (numeric, mM)}
\item{inhibition_actual}{Inhibitor condition (e.g. "no_inhibition", "competitive", "noncompetitive", "uncompetitive")}
}
}
\source{
Generated for testing
}
\usage{
enzyme_properties
}
\description{
A dataset of enzyme kinetic and inhibition parameters used in simulations
Substrate set is alcohol by default
}
\keyword{datasets}
