% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{estimateKmVmax}
\alias{estimateKmVmax}
\title{Calculate estimated \code{Km} and \code{Vmax} values
Fits Michaelis-Menten equation to substrate concentrations and reaction rates using nonlinear least squares regression}
\usage{
estimateKmVmax(df)
}
\value{
Estimated values for gradient when called in \code{processReactionParameters}. User does not need to interact.
}
\description{
Calculate estimated \code{Km} and \code{Vmax} values
Fits Michaelis-Menten equation to substrate concentrations and reaction rates using nonlinear least squares regression
}
\details{
Uses \code{\link[stats]{nls}} to estimate Km and Vmax.
}
\seealso{
\code{\link{processReactionParameters}} \code{\link[stats]{nls}}
}
