% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStudentSummaryPDF.R
\name{generateStudentSummaryPDF}
\alias{generateStudentSummaryPDF}
\title{Generate student summary PDFs
Generates individual and/or combined summary PDF files for each student}
\usage{
generateStudentSummaryPDF(individual = individual, combined = combined)
}
\arguments{
\item{individual}{Logical. if TRUE (default), then maudr will generate individual student PDFs}

\item{combined}{Logical. if TRUE (default), then maudr will generate a single combined PDF of all students}
}
\value{
\code{.PDF} files in the \verb{/Answer_files} subdir of the latest dir under \verb{output/}. Is called as a part of the \code{runMaudr} pipeline.
}
\description{
Generate student summary PDFs
Generates individual and/or combined summary PDF files for each student
}
\details{
Internally calls logic from \code{plotAbsVsTime}, \code{plotLB}, and \code{plotMM} to generate summary PDFs for each student,
combines these plots and the output from \code{processReactionParameters}, dependent on what parameters the user inputs.
}
\seealso{
\code{\link{runMaudr}} \code{\link{plotAbsVsTime}} \code{\link{plotLB}} \code{\link{plotMM}}
}
