% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialiseMaudrData.R
\name{initialiseMaudrData}
\alias{initialiseMaudrData}
\title{Initialise the maudr data directory with sample data
Creates a \verb{data/} folder in the working directory (if it doesn't exist)
and copies in default \code{.rda} files from the maudr package if the user hasn't already
added their own .xlsx or .rda files.}
\usage{
initialiseMaudrData()
}
\value{
No return value. Called to ensure \verb{data/} folder exists, and loads in provided samples if \verb{data/} is empty.
}
\description{
Initialise the maudr data directory with sample data
Creates a \verb{data/} folder in the working directory (if it doesn't exist)
and copies in default \code{.rda} files from the maudr package if the user hasn't already
added their own .xlsx or .rda files.
}
\details{
Checks for three core data files: \code{enzyme_properties.rda}, \code{student_names.rda}, and \code{variable_storage.rda}.
Looks for either user provided \code{.xlsx} or \code{.rda} files within \verb{data/} directory. If neither are found, built-in samples
are written to \verb{data/}.
Function is called as a part of the \code{runMaudr()} pipeline but can be run manually to reinitialise data.
}
\seealso{
\code{\link{runMaudr}}
}
