% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLB}
\alias{plotLB}
\title{Plot Lineweaver-Burk for each student}
\usage{
plotLB()
}
\value{
Used internally by \code{generateStudentSummaryPDF}, list of plot objects
}
\description{
Plot Lineweaver-Burk for each student
}
\details{
Generates a Lineweaver-Burk (double recip.) plot for each student using the \code{students_rxn_params.rda} output of \code{assignReactionConditions}
Both inhibition types are stacked and differentiated by colour (colour is static)
}
\seealso{
\code{\link{plotAbsVsTime}}, \code{\link{plotMM}}, \code{\link{generateStudentSummaryPDF}}
}
