% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMM}
\alias{plotMM}
\title{Plot Michaelis-Menten curves for each student}
\usage{
plotMM()
}
\value{
Used internally by \code{generateStudentSummaryPDF}
}
\description{
Plot Michaelis-Menten curves for each student
}
\details{
Generates a Michaelis-Menten plot for each student using the \code{students_rxn_params.rda} output of \code{assignReactionConditions}
Both inhibition types are stacked and differentiated by colour (colour is static)
}
\seealso{
\code{\link{plotAbsVsTime}}, \code{\link{plotLB}}, \code{\link{generateStudentSummaryPDF}}
}
