% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processReactionParameters.R
\name{processReactionParameters}
\alias{processReactionParameters}
\title{Process and estimates kinetic parameters for students
Locates latest timestamped directory in \verb{\\output}, calculate reaction rates for all substrate concentrations, fits Michaelis-Menten
parameters per student and inhibition type}
\usage{
processReactionParameters()
}
\value{
Processed data frame with estimated kinetic parameters for each student and substrate. Saves to \verb{\\variables} directory but is irrelevant to average user.
}
\description{
Process and estimates kinetic parameters for students
Locates latest timestamped directory in \verb{\\output}, calculate reaction rates for all substrate concentrations, fits Michaelis-Menten
parameters per student and inhibition type
}
\details{
Loads student parameters and calculates reaction rate with \code{calculateReactionRate}, fits to Michaelis-Menten curves using \code{estimateKmVmax},
stores the results as \code{processed_data.rda} in the output directory for downstream use.
}
\seealso{
\code{\link{runMaudr}} \code{\link{calculateReactionRate}}, \code{\link{estimateKmVmax}}, \code{\link{generateAbsVsTimeData}}, \code{\link{generateStudentSummaryPDF}}
}
