% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{student_names}
\alias{student_names}
\title{Example student names}
\format{
A data frame with 3 variables:
\describe{
\item{student_no}{Student number, formatted as a string (e.g. "u123444")}
\item{first_name}{First name of student}
\item{surname}{Surname of student}
}
}
\source{
Generated for testing
}
\usage{
student_names
}
\description{
Small dataset containing sample student IDs and names for use in \code{maudr}
}
\keyword{datasets}
