% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{variable_storage}
\alias{variable_storage}
\title{Example variable storage}
\format{
A data frame with 5 variables:
\describe{
\item{time_sec}{Reaction time (in seconds)}
\item{substrate_conc_mM}{Substrate concentration (in mM)}
\item{extinction_coeff}{Extinction coefficient (numeric, 6220 M\eqn{^{-1}} cm\eqn{^{-1}})}
\item{cuvette_volume_l}{Cuvette volume (in liters, numeric)}
\item{enzyme_volume_ml}{Enzyme volume (in milliliters, numeric)}
}
}
\source{
Generated for testing
}
\usage{
variable_storage
}
\description{
A dataset of experimental constants for calculating reaction conditions
The extinction coefficient provided is for alcohol dehydrogenase (ADH), same with the enzyme substrate set
}
\keyword{datasets}
